refactored data_source template into its own header
